// apps/vuttik-pos-web/src/pos/types.ts
export interface Commission {
  id: string;
  employee_id: string;
  employee_nombre: string;
  sale_id: string;
  product_id: string;
  product_nombre: string;
  monto: number;
  porcentaje: number;
  fecha: any;
  estado: 'pendiente' | 'retirada';
}